backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.1
pydantic-settings>=2.1.0
slowapi>=0.1.9
python-multipart>=0.0.6
redis>=5.0.0
celery>=5.3.0
httpx>=0.26.0
beautifulsoup4>=4.12.0
